lifts raise_this test utility to central module
