+++ tests/recruitment/agency/__init__.py
@@ -23,6 +23,13 @@ def client(client_type: str, region_name: str):
     return get_session().create_client(client_type, region_name=region_name)
 
 
+def raise_this(**kwargs):
+    exception = kwargs['exception']
+    def will_raise(**kwargs):
+        raise exception
+    return will_raise
+
+
 @contextmanager
 def uncloseable(buffer: Buffer):
     """Context manager which makes the buffer's close operation a no-op"""
